refactor(pr07): promote sheet plumbing to settings/shared/sheet (G5f.6)
git mv src/settings/core/sheet -> src/settings/shared/sheet. After narrowing core/ to hidden settings, the generic settings sheet plumbing (GenericSettingsApp.vue, SettingsTable, settingsStore) no longer belongs there. The new shared/ folder is named to accommodate future cross-cutting helpers.

Update all importers (~9 files) from @settings/core/sheet/* to @settings/shared/sheet/*. Drop GenericSettingsApp/SettingsStoreSymbol/useSettingsStore re-exports from core/index.mts (now re-exported via shared/index.mts which settings/index.mts pulls in). Repoint GenericSettingsApp.vue's SettingsSection type import to @settings/core/_types.mjs.
