+++ src/entities/items/components/Physical/sheet/PhysicalItemStore.mts
@@ -7,11 +7,11 @@ import type { SecretType } from '@effects/secret/Secret.mjs';
 import { secretEffectType } from '@effects/secret/secretEffectType.mjs';
 import type { ItemDocumentActions, ItemDocumentGetters, ItemSheetStore, ItemSheetStoreUtils } from '@items/baseItem/index.mjs';
 import { physicalItemEffectsTab, type PhysicalItemLike } from '@items/components/Physical/index.mjs';
-import type { SettingsStore } from '@settings/core/sheet/settingsStore.mjs';
 import { PriceData } from '@settings/currency/index.mjs';
 import type { DamageReductionTypesConfig } from '@settings/index.mjs';
 import { GAME_RULES_KEYS, SettingsStoreSymbol } from '@settings/index.mjs';
 import { SYSTEM_ID } from '@settings/shared.mjs';
+import type { SettingsStore } from '@settings/shared/sheet/settingsStore.mjs';
 import type { MultiSelectOption } from '@vc/Fields/index.mjs';
 import type { VueApplicationContext } from '@vueApps/index.mjs';
 import type { ComputedRef } from 'vue';
