+++ src/entities/items/components/Physical/sheet/components/ItemWeight.vue
@@ -16,8 +16,8 @@
 <script setup lang="ts">
   import { DocumentSheetStoreSymbol } from '@ec/CoreMixin/index.mjs';
   import type { PhysicalDocumentStore } from '@items/components/Physical/index.mjs';
-  import type { SettingsStore } from '@settings/core/sheet/index.mjs';
-  import { SettingsStoreSymbol } from '@settings/core/sheet/index.mjs';
+  import type { SettingsStore } from '@settings/shared/sheet/index.mjs';
+  import { SettingsStoreSymbol } from '@settings/shared/sheet/index.mjs';
   import { NumberFormGroup } from '@vc/Fields/index.mjs';
   import { inject, useSlots } from 'vue';
