+++ src/settings/core/index.mts
@@ -5,4 +5,3 @@
 export type { FieldChoice, SettingField, SettingsSection } from './_types.mjs';
 export { CORE_KEYS } from './constants.mjs';
 export { registerCoreSettings } from './registration.mjs';
-export { GenericSettingsApp, SettingsStoreSymbol, useSettingsStore } from './sheet/index.mjs';
