+++ src/settings/display/sheet/DisplaySettingsConfig.mts
@@ -8,8 +8,9 @@ import { useVueSettingsMixin } from '@vueApps/index.mjs';
 import type { App, Component } from 'vue';
 import { createApp } from 'vue';
 
-import { GenericSettingsApp, type SettingsSection } from '../../core/index.mjs';
+import type { SettingsSection } from '../../core/index.mjs';
 import { SYSTEM_ID } from '../../shared.mjs';
+import { GenericSettingsApp } from '../../shared/sheet/index.mjs';
 import { DISPLAY_KEYS } from '../constants.mjs';
 
 // Foundry global UI reference
