+++ src/settings/shared/index.mts
@@ -1,0 +1,9 @@
+/**
+ * Shared settings infrastructure
+ *
+ * Cross-cutting helpers, components, and stores used by multiple settings
+ * categories (and elsewhere in the system). Currently re-exports the generic
+ * settings sheet plumbing; intended to grow as more shared concerns emerge.
+ */
+
+export * from './sheet/index.mjs';
